move HomePage.tsx page to LandingSection component
